syndwarsfx: Provide 32-bit ASM input regs in create_electric_strand()
A bug was intrduced in the patch:
"syndwarsfx: Remade init_taser()"

The bug - parameters treated as 32-bit in assembly were provided as
16-bit, making negative values positive.

The bug was visible on Orbital Station missions, ie. `-m 0,88`.

+++ src/thing.c
@@ -2017,7 +2017,7 @@ struct SimpleThing *create_electric_strand(MapCoord x, MapCoord y, MapCoord z,
       "push 0(%5)\n"
       "call ASM_create_electric_strand\n"
         : "=r" (ret)
-        : "a" (x), "d" (y), "b" (z), "c" (x2), "S" (stkargs)
+        : "a" ((s32)x), "d" ((s32)y), "b" ((s32)z), "c" ((s32)x2), "S" (stkargs)
         : "cc", "memory");
     return ret;
 #endif
